Pull the CSS into the iframe project.
A couple unfortunate caveats:

* I had to leave the images and fonts back in the rails project to 
avoid a really complex asset_path issue. Basically if CSS references an
asset by name, and we're using rails-lineman, then our lineman css 
cannot know the substitution it needs to make to add a font or image
asset fingerprint.
* As a result, the iframe is "mostly almost totally separate" but in 
practice will never be sensibly worked on in a totally separate context
from the backing rails project. This is a moral defeat, but it still
gives us the toolset we need to start trimming what we do control.