Serve up a mangled HTML response for facebot
Facebook’s scraper does two things that defeats our shortlink mechanism:

1. It follows redirects and discards shorturls along the way, meaning
a user would be sharing a raw publisher URL and not the impaq.me 
shortlink
2. When a link-rel-canonical is set or when a meta-og:url is present in
the HTML, facebook will scrape that URL instead of the shortlink. 

This commit defeats #1 by fetching the HTML ourselves and proxying the
response to users. It defeats #2 by mangling the HTML response with 
regex. Yes, really.