Allow campaign selection by SDK demo page
An outstanding annoyance is that if you want to test the widget for a
particular campaign you have to go dig to set the current domain of
the demo page (usually localhost) to whatever campaign you're testing
and remove it from the others. This commit lets you just type that in,
tab out, refresh page, and force the page to load the widget for X 
campaign.

Because this is something we would not want to expose in production,
I hid all of the duck-punching behind an initializer that only runs
in development.