Add logical deletion of campaigns
To delete a campaign go to its edit page:

/campaigns/edit/11

And click "delete campaign" (a prompt will warn you to confirm).


If you erroneously delete a campaign, fear not, as it's merely being
hidden from the app. To undelete it, you'll just need to get into a
rails console and run `Campaign.undelete!(11)`

Fixes #118